Implementation of priority sectors (#240)
* Create highlight sectors

* Insert prioritized sectors in db

* Modify budget

* Reorganize Sector Id code

* Add System.Diagnostics

* Change treeIndexes to uint

* Add better types to TreeIndex to SectorId map

* Order nodes by treeindex only once

* Remove StidTagMapper stuff (Moving it to a its own branch)

* Simplify SequentialIdGenerator

* Update tests

* More test

* Expand to max id 2^24 for real

* Simplify TreeIndexGenerator

* Add more comments

* Rename dict and move logigng

* Adding tests for PrioritySectorSplitter

* Add priority and discipline to protobuf

* Add tests for PrioritySplittingUtils

* Fix small issue in CameraPositioningTests

* Rename filename used for prioritized sectors

* Add/fix a couple of comments

* Log how many primitives was considered outliers

* Remove TODO and create story + create a method for post process db manipulation

---------

+++ CadRevealComposer/CadRevealNode.cs
@@ -64,7 +64,7 @@ public record BoundingBox([property: ProtoMember(1)] Vector3 Min, [property: Pro
 
 public class CadRevealNode
 {
-    public required ulong TreeIndex { get; init; }
+    public required uint TreeIndex { get; init; }
     public required string Name { get; init; }
 
     // TODO support Store, Model, File and maybe not RVM
